Remove license headers from source files
This is not required, this is just some un-needed visual
garbage. The license is already described in LICENSE.txt
and mentioned in the readme and documentation. This will
make the maintainance easier when we have to update copyright
years.

This will also fix most of the PEP8 issues on travis.