Kill server process to avoid any downtime
on windows the process sometimes takes more than 5s to close
which will cause a visible lag when used in a QTabWidget. It's better
to kill it than having a lag, the consequence is that an error message
will be printed in stderr: QProcess: Destroyed while process is still running.