CodeCompletion: fix performace issue due to recent change in sync system
Before this commint, the system only display the result of the latest completion request which is always reset when you type too fast which prevent the popup from being shown. The original issue was that an older request complete after the latest. The solution is to discard any request whose request id is lower than the last complete request. That way we don't show outdated request but show early request which gives the sensation of a better responsibity (very subtile).