Simplify context menu
Don't show standard actions (undo, redo, copy, paste,...) by default

The create_default_actions now only relate to those standard actions, if set to True they will get created. In any case, non standard actions will get created (goto line and duplicate line).

To prevent the context menu from showing up, you just need to set the new property `show_context_menu` to False.

The goal of this change is to reduce the context menu to only the actions that the user might be looking for (we never use the copy/paste menu but the shortcuts so it is not usefull to display them in the context menu).