Fix slowness in search when there are thousands of occurences to highlight
The search itself is fast, it is able to find more than 15k occurences of a character in a document of 13k lines, but the highlighting
is slow. To workaround this limitation, there is now maximum number of highlighted occurences. Default is 500 which should be more than enough
for normal usage. The occurences that are not highlighted can be reached by cycling though the results, this is just a display limitation