Fix a another potential cause of memory leak
and add some log message for signaling the deletion of the editor widget and its modes.

SplittableTabWidget: add a warning if a potential memory leak is detected when an editor widget has been closed (if objgraph has been installed it will output a backrefs graph showing the objects that are still holding a reference to the editor).