Fix stange behaviour when resizing a dock widget on Plasma 5
This happens in hackedit when an editor is open and user try to resize one of the left/right dock widgets. This is especially visible on Plasma 5 but also happen on Ubuntu 14.04 when the system has high CPU usage.

The bug that the removed code was meant to fix was actually a bug in SplittableTabWidget (fixed a long time ago).